requirements.txt
# Core FPL API
requests>=2.31.0
pyyaml>=6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Database (THIS IS WHAT'S MISSING!)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
supabase>=2.0.0

# Optimization
pulp>=2.7.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
tabulate==0.9.0

# FPL Sentinel (Twitter Scraper & Sentiment Analysis)
selenium>=4.15.0
undetected-chromedriver>=3.5.0
textblob>=0.17.1
vaderSentiment>=3.3.2

# Dashboard API
fastapi>=0.104.0
uvicorn>=0.24.0

